requirements.txt
feedparser
pydantic-settings
python-dotenv
python-dateutil
rapidfuzz
sentence-transformers
scikit-learn
numpy
requests-oauthlib
beautifulsoup4
PyYAML
